refactor[cm]: reorganize ICMBC classes
- move all ICMBC logic to ICMBC folder
- move Edge and SubGraphPair classes to separate files
- move some CaDETMember specific methods to CaDETMember
- optimize searching for member's directly and indirectly accessed fields
